Crate conrod_winit

A function for converting a winit::Event to a conrod::event::Input.

Modules

macros
A collection of macros for generating code related to winit+conrod interop.

v020

v021

v022

v023

Macros

conversion_fns
Generate a set of conversion functions for converting between types of the crate's versions of winit and conrod_core.

convert_event
A macro for converting a winit::Event to a conrod_core::event::Input.

convert_key
Maps winit's key to a conrod Key.

convert_mouse_button
Maps winit's mouse button to conrod's mouse button.

convert_mouse_cursor
Convert a given conrod mouse cursor to the corresponding winit cursor type.

convert_window_event
A macro for converting a winit::WindowEvent to a Option<conrod_core::event::Input>.

v020_conversion_fns

v020_convert_event
A macro for converting a winit::Event to a conrod_core::event::Input.

v020_convert_key

v020_convert_mouse_button
Maps winit's mouse button to conrod's mouse button.

v020_convert_mouse_cursor
Convert a given conrod mouse cursor to the corresponding winit cursor type.

v020_convert_window_event
A macro for converting a winit::WindowEvent to a Option<conrod_core::event::Input>.

v021_conversion_fns

v021_convert_event
A macro for converting a winit::Event to a conrod_core::event::Input.

v021_convert_key

v021_convert_mouse_button
Maps winit's mouse button to conrod's mouse button.

v021_convert_mouse_cursor
Convert a given conrod mouse cursor to the corresponding winit cursor type.

v021_convert_window_event
A macro for converting a winit::WindowEvent to a Option<conrod_core::event::Input>.

v022_conversion_fns

v022_convert_event
A macro for converting a winit::Event to a conrod_core::event::Input.

v022_convert_key

v022_convert_mouse_button
Maps winit's mouse button to conrod's mouse button.

v022_convert_mouse_cursor
Convert a given conrod mouse cursor to the corresponding winit cursor type.

v022_convert_window_event
A macro for converting a winit::WindowEvent to a Option<conrod_core::event::Input>.

v023_conversion_fns

v023_convert_event
A macro for converting a winit::Event to a conrod_core::event::Input.

v023_convert_key

v023_convert_mouse_button
Maps winit's mouse button to conrod's mouse button.

v023_convert_mouse_cursor
Convert a given conrod mouse cursor to the corresponding winit cursor type.

v023_convert_window_event
A macro for converting a winit::WindowEvent to a Option<conrod_core::event::Input>.

Traits

WinitWindow
Types that have access to a winit::Window and can provide the necessary dimensions and hidpi factor for converting winit::Events to conrod::event::Input, as well as set the mouse cursor.