Macro conrod_winit::v021_convert_event

macro_rules! v021_convert_event {
    ($event:expr, $window:expr) => { ... };
}

A macro for converting a winit::Event to a conrod_core::event::Input.

Expects a winit::Event and a reference to a window implementing WinitWindow. Returns an Option<conrod_core::event::Input>.

Invocations of this macro require that a version of the winit and conrod_core crates are available in the crate root.